refactor(gsd): narrow merge worktree test seams
