+++ src/resources/extensions/gsd/auto-worktree-entry.ts
@@ -1,0 +1,6 @@
+// gsd-pi — Narrow auto-worktree entry/query seam.
+//
+// Keeps callers that only enter or identify auto-worktrees off the legacy
+// auto-worktree compatibility barrel while extraction continues.
+
+export { enterAutoWorktree, isInAutoWorktree } from "./auto-worktree.js";
