+++ src/resources/extensions/gsd/auto-worktree-session-registry.ts
@@ -1,0 +1,10 @@
+// gsd-pi — Narrow auto-worktree session registry seam.
+//
+// Provides registry state/reset helpers without exposing the whole legacy
+// auto-worktree compatibility barrel to tests.
+
+export {
+  getActiveAutoWorktreeContext,
+  getAutoWorktreeOriginalBase,
+  _resetAutoWorktreeOriginalBaseForTests,
+} from "./auto-worktree.js";
